Set partition with math problem
There is a way to solve this problem very clumsy with
a bunch of if statements. Breaking appart each of the possible
cases and trying to match a logic for each one.

But here thinking with math really helps. Going for the upper
bound of the amount of taxis, n to the lower bound - the number
of groups with 4. And them closing this bound.

Signed-off-by: Luã Santilli <luasantilli@gmail.com>